libhwbinder: Update zeroMemory for HIDL transactions
Due to compiler optimizations, there are cases where memset is
optimized out of the actual compiled artifact if the compiler
knows that memory cell won't be access by the process. For
security reasons this isn't ideal, this change introduces
memset_explicit and an assembly instruction to match memset_explicit
as a fallback.

Bug: 416248344
Test: hidl_test
Change-Id: Iacd616b423fb4f92d36c13c127523d44233e0fb0