libhwbinder: Await hwservicemanager.disabled first
Currently libhwbinder waits for hwservicemanager.ready with timeout of 1
second in a loop. However, On devices with hwservicemanager disabled,
this property is never set, and that causes a race that forces
libhwbinder to wait for the timeout. This can be fixed by letting
libhwbinder wait for hwservicemanager.disabled first, but
hwservicemanager.disabled was never set on devices with hwservicemanager
enabled.

Now we are introducing a change to always set hwservicemanager.disabled
to true or false on all devices. So, libhwbinder can wait for
hwservicemanager.disabled first. By this change, libhwbinder is no
longer forced to be blocked for 1 second.

Bug: 419725480
Test: manual
Change-Id: I89aaa3a42e355ae3c6a2a0b2ad6747b3709c2ead