libhwbinder: wait for hwservicemanager.ready to be initialized
This is a reland of cc6e8e6aa1a6e36660cc1cf5250b0ed74b535d2b.

Currently libhwbinder waits for hwservicemanager.ready to be true with
1-sec timeout in a loop. However, On devices with hwservicemanager
disabled, .ready is never set, and that causes a race that forces
libhwbinder to wait for the timeout. This unnecessarily slows down the
boot time.

Now we are introducing a change to always set hwservicemanager.disabled
to false verbosely even on devices where hwservicemanager is enabled.
So, libhwbinder can wait for hwservicemanager.disabled first. By this
change, libhwbinder is no longer forced to be blocked for 1 second.

There is no guarantee libhwbinder and hwservicemanager are updated
together. Actually, older hwservicemanager can run with newer
libhwbinder and vice versa. So, we set
`hwservicemanager.always_sets_disabled` system property to detect the
new behavior.

By this change, the boot time of some devices will be improved by around
300ms.

Bug: 419725480
Test: CF boots. Confirmed that .disabled is checked by additional log
Test: Confirmed no timeout messages appear in logcat on CF
Test: CF boots with ToT hwservicemanager as well
Change-Id: Iee2e8ed034ac0fd7bec1bdffd66a70c79e1a4990